Fix timing issues for RS485 master-slave communication.
Three fixes:

1. Send a dummy 0xff as the first byte of transmission. This allows
the receivers UART state machine to sync up, if it happens to think it
is in the middle of a character when we start transmitting.

2. Introduce a start-of-frame character ('?' for master request), and ignore any characters received before the start-of-frame character.

3. Properly wait for the last character to be transmitted before
switching off transmit mode, to avoid dropping any parts of the last
character(s) to be sent. (This turns out to require some care due to
the way Atmega328 handles the TXC "transmission complete" bit).